refactor: step2 에서 구현한 클래스 페이지 이동
