+++ src/main/java/study/step2/Calculator.java
@@ -1,4 +1,4 @@
-package study;
+package study.step2;
 
 import java.util.ArrayList;
 import java.util.List;
