+++ src/main/java/study/step2/Operator.java
@@ -1,4 +1,4 @@
-package study;
+package study.step2;
 
 import java.util.Arrays;
 import java.util.function.BiFunction;
