+++ src/test/java/study/step2/CalculatorTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.step2;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
